+++ 33ma3/src/main/java/softeer/be33ma3/repository/post/PostCustomRepository.java
@@ -1,4 +1,4 @@
-package softeer.be33ma3.repository;
+package softeer.be33ma3.repository.post;
 
 import softeer.be33ma3.domain.Post;
