+++ 33ma3/src/main/java/softeer/be33ma3/repository/post/PostRepository.java
@@ -1,7 +1,8 @@
-package softeer.be33ma3.repository;
+package softeer.be33ma3.repository.post;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import softeer.be33ma3.domain.Post;
+import softeer.be33ma3.repository.post.PostCustomRepository;
 
 import java.util.List;
