+++ 33ma3/src/test/java/softeer/be33ma3/service/ChatServiceTest.java
@@ -12,6 +12,7 @@ import softeer.be33ma3.domain.Member;
 import softeer.be33ma3.domain.Post;
 import softeer.be33ma3.dto.request.PostCreateDto;
 import softeer.be33ma3.repository.*;
+import softeer.be33ma3.repository.post.PostRepository;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -24,7 +25,8 @@ class ChatServiceTest {
     @Autowired ChatService chatService;
     @Autowired MemberRepository memberRepository;
     @Autowired ChatRoomRepository chatRoomRepository;
-    @Autowired PostRepository postRepository;
+    @Autowired
+    PostRepository postRepository;
     @Autowired ChatMessageRepository chatMessageRepository;
     @Autowired AlertRepository alertRepository;
