+++ 33ma3/src/test/java/softeer/be33ma3/service/OfferServiceTest.java
@@ -15,7 +15,7 @@ import softeer.be33ma3.exception.BusinessException;
 import softeer.be33ma3.repository.CenterRepository;
 import softeer.be33ma3.repository.MemberRepository;
 import softeer.be33ma3.repository.OfferRepository;
-import softeer.be33ma3.repository.PostRepository;
+import softeer.be33ma3.repository.post.PostRepository;
 
 import java.util.ArrayList;
 import java.util.List;
